+++ src/main/java/org/springframework/samples/petclinic/web/PdfController.java
@@ -5,8 +5,11 @@ import javax.validation.Valid;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.samples.petclinic.model.Pdf;
+import org.springframework.samples.petclinic.model.Usuario;
 import org.springframework.samples.petclinic.service.PdfService;
 import org.springframework.samples.petclinic.service.UsuarioService;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.ModelMap;
 import org.springframework.validation.BindingResult;
@@ -21,6 +24,9 @@ public class PdfController {
 	public static final String PDFs_FORM = "pdfs/createOrUpdatePdfsForm";
 	public static final String PDFs_LISTING = "pdfs/PdfsListing";
 	public static final String PDFs_VISUALIZE = "pdfs/PdfVisualize";
+	public static final String MEJORAR_CUENTA = "usuarios/mejorarCuenta";
+	public static final String LOGIN = "login";
+	public static final String ERROR = "";
 
 	
 	private final PdfService pdfService;
@@ -35,7 +41,11 @@ public class PdfController {
 	@GetMapping
 	public String listPdfs(ModelMap model) {
 		model.addAttribute("pdfs", pdfService.findAll());
-		model.addAttribute("usuarios", usuarioService.findAll());
+		String authority = "";
+		if (AuthController.isAuthenticated()) {
+			authority = AuthController.highestLevel();
+		}
+		model.addAttribute("authority", authority);
 		return PDFs_LISTING;
 	}
 	
@@ -49,6 +59,12 @@ public class PdfController {
 	@GetMapping("/{id}/edit")
 	public String editPdf(@PathVariable("id") int id, ModelMap model) {
 		Pdf pdf = pdfService.findById(id);
+		Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+		String username = authentication.getName();
+		Usuario usuarioLoggeado = usuarioService.findByUsername(username);
+		if (!pdf.getUsuario().equals(usuarioLoggeado) && !AuthController.isAdmin()) {
+			return "redirect:/" + ERROR;
+		}
 		model.addAttribute("pdf", pdf);
 		return PDFs_FORM;
 	}
@@ -71,13 +87,25 @@ public class PdfController {
 	@GetMapping("/{id}/delete")
 	public String deletePdf(@PathVariable("id") int id,ModelMap model) {
 		Pdf pdf=pdfService.findById(id);
+		Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+		String username = authentication.getName();
+		Usuario usuarioLoggeado = usuarioService.findByUsername(username);
+		if (!pdf.getUsuario().equals(usuarioLoggeado) && !AuthController.isAdmin()) {
+			return "redirect:/" + ERROR;
+		}
 		pdfService.delete(pdf);
-		model.addAttribute("message","The thread was deleted successfully!");
+		model.addAttribute("message","El documento ha sido eliminado con éxito.");
 		return listPdfs(model);
 	}
 	
 	@GetMapping("/new")
 	public String editNewDisease(ModelMap model) {
+		if (!AuthController.isAuthenticated()) {
+			return "redirect:/" + LOGIN;
+		}
+		if (!AuthController.hasPaid()) {
+			return "redirect:/" + MEJORAR_CUENTA;
+		}
 		model.addAttribute("pdf",new Pdf());
 		return PDFs_FORM;
 	}
@@ -88,6 +116,10 @@ public class PdfController {
 			model.addAttribute("message", "Documento inválido.");
 			return PDFs_FORM;
 		}else {
+			Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+			String username = authentication.getName();
+			Usuario usuarioLoggeado = usuarioService.findByUsername(username);
+			pdf.setUsuario(usuarioLoggeado);
 			pdfService.save(pdf);
 			model.addAttribute("message", "The thread was created successfully!");			
 			return listPdfs(model);
